requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
openai==1.58.1
httpx==0.28.1
python-dotenv==1.0.1
streamlit==1.41.1
requests==2.32.3
prometheus-client==0.21.1
